requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.3
deepl==1.18.0
defusedxml==0.8.0rc2
distlib==0.3.8
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==3.4.0
django-dbbackup==4.2.1
django-livereload-server==0.5.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dropbox==12.0.2
exceptiongroup==1.2.2
filelock==3.15.4
git-filter-repo==2.45.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
jiter==0.7.0
lxml==5.2.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
openai==1.54.3
packaging==24.1
pillow==11.0.0
pipenv==2024.0.1
platformdirs==4.2.2
ply==3.11
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
stone==3.3.1
tornado==6.4.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
whitenoise==6.7.0
